backend/pyproject.toml
[tool.poetry]
name = "crud"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.24"
psycopg2 = "^2.9.9"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.6.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
